feat: add index.php and main.md for lab 2

+++ lab2/index.php
@@ -4,7 +4,7 @@ echo "Привет, мир!";
 echo "Hello, World with echo!";
 print "Hello, World with print!";
 
-$days = (int) 228;
+$days = 228;
 $message = "Все возвращаются на работу!";
 
 echo "Через " . $days . " дней " . $message . "<br />";
